==> notebooks/01_eda_diabetes_prediction.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline

df = pd.read_csv('../data/raw/diabetes_prediction_dataset.csv')


df.head()

df.info()

df.describe()

df.corr(numeric_only=True)

sns.heatmap(df.corr(numeric_only=True), annot=True, cmap='coolwarm', fmt='.2f')

sns.pairplot(df)

df[df['age'] < 1].count()

df[df['bmi'] > 80].count()

sns.histplot(df['age'], bins=50, kde=True)


sns.histplot(df['bmi'], bins=50, kde=True)

df_clean = df[(df['age'] >= 1) & (df['bmi'] <= 80)]
print(f"Rows before cleaning: {len(df)}")
print(f"Rows after cleaning: {len(df_clean)}")


sns.histplot(df_clean['blood_glucose_level'], bins=50, kde=True)

sns.histplot(df_clean['HbA1c_level'], bins=50, kde=True)

df_clean['diabetes'].value_counts()

df_clean.to_csv('../data/processed/diabetes_prediction_clean.csv', index=False)
